+++ src/redisearch_rs/trie_rs/src/trie_map/iter/automaton/wildcard/atoms.rs
@@ -7,185 +7,11 @@
  * GNU Affero General Public License v3 (AGPLv3).
 */
 
-//! Low-level building blocks for the wildcard NFA: the [`Atom`] enum, the
-//! [`NfaBitSet`] trait, and the `flatten` helper that turns a parsed
-//! pattern into an atom slice. See the [parent module
-//! doc](super) for the NFA primer this fits into.
-//!
-//! The bitset is monomorphised twice — once for `u64` (≤ 63 atoms) and
-//! once for `u128` (≤ 127 atoms). Each `NfaBitSet` impl is a few
-//! single-register operations, so every method on
-//! [`super::nfa::WildcardNfa<S>`] inlines into a tight bitwise hot loop.
-//! [`super::WildcardBackend::for_pattern`] picks the width; patterns
-//! past 127 atoms skip the NFA entirely and route to
-//! [`crate::trie_map::iter::WildcardIter`].
+//! Low-level building blocks for the wildcard NFA: the [`Atom`] enum and
+//! the `flatten` helper that turns a parsed pattern into an atom slice.
+//! See the [parent module doc](super) for the NFA primer this fits into.
 
 use rqe_wildcard::{Token, WildcardPattern};
-
-/// Bitmask of active NFA positions, with bit *i* set iff position *i*
-/// (in the sense documented on the [parent module](super)) is in the
-/// current active set. The full set must fit in the type — implementers
-/// reserve capacity for positions `0..=n_atoms` (the `+1` is the accept
-/// position).
-///
-/// Methods mirror standard set operations. Width-specific implementations
-/// are provided for `u64` and `u128`.
-pub trait NfaBitSet: Clone + Eq {
-    /// Number of distinct positions this bitset can hold. A pattern with
-    /// `n` atoms uses positions `0..=n` (accept included), so the backend
-    /// is only sound when `n < CAPACITY`. [`super::nfa::WildcardNfa::compile`]
-    /// enforces this; the dispatcher in [`super::WildcardSpecializedIter`]
-    /// picks the narrowest width that fits.
-    const CAPACITY: usize;
-
-    /// Empty set, sized for positions `0..=n_atoms`.
-    fn empty(n_atoms: usize) -> Self;
-
-    /// Set containing only `pos`, sized for positions `0..=n_atoms`.
-    fn singleton(n_atoms: usize, pos: usize) -> Self;
-
-    /// Reset every bit to zero. Used to recycle a scratch set across
-    /// successive transitions inside [`super::nfa::nfa_step_into`].
-    fn clear(&mut self);
-
-    /// Add `pos` to the set.
-    fn insert(&mut self, pos: usize);
-
-    /// Whether `pos` is in the set.
-    fn contains(&self, pos: usize) -> bool;
-
-    /// Whether the set is empty (the NFA has died and the subtree can
-    /// be pruned).
-    fn is_empty(&self) -> bool;
-
-    /// `self |= other` (set union). Both operands must be the same width.
-    fn union_in_place(&mut self, other: &Self);
-
-    /// Position of the lone set bit.
-    ///
-    /// If the set has more than one element, it returns the position
-    /// of the highest set bit.
-    fn singleton_pos(&self) -> usize;
-
-    /// Iterate active positions in ascending order.
-    fn iter(&self) -> impl Iterator<Item = usize> + '_;
-}
-
-impl NfaBitSet for u64 {
-    const CAPACITY: usize = 64;
-
-    #[inline]
-    fn empty(_: usize) -> Self {
-        0
-    }
-    #[inline]
-    fn singleton(_: usize, pos: usize) -> Self {
-        debug_assert!(pos < 64);
-        1u64 << pos
-    }
-    #[inline]
-    fn clear(&mut self) {
-        *self = 0;
-    }
-    #[inline]
-    fn insert(&mut self, pos: usize) {
-        debug_assert!(pos < 64);
-        *self |= 1u64 << pos;
-    }
-    #[inline]
-    fn contains(&self, pos: usize) -> bool {
-        debug_assert!(pos < 64);
-        (*self >> pos) & 1 == 1
-    }
-    #[inline]
-    fn is_empty(&self) -> bool {
-        *self == 0
-    }
-    #[inline]
-    fn union_in_place(&mut self, other: &Self) {
-        *self |= *other;
-    }
-    #[inline]
-    fn singleton_pos(&self) -> usize {
-        debug_assert_eq!(
-            self.count_ones(),
-            1,
-            "Invoked 'singleton_pos' on a set that contains more than one element."
-        );
-        self.trailing_zeros() as usize
-    }
-    #[inline]
-    fn iter(&self) -> impl Iterator<Item = usize> + '_ {
-        let mut bits = *self;
-        std::iter::from_fn(move || {
-            if bits == 0 {
-                None
-            } else {
-                let pos = bits.trailing_zeros() as usize;
-                bits &= bits - 1;
-                Some(pos)
-            }
-        })
-    }
-}
-
-impl NfaBitSet for u128 {
-    const CAPACITY: usize = 128;
-
-    #[inline]
-    fn empty(_: usize) -> Self {
-        0
-    }
-    #[inline]
-    fn singleton(_: usize, pos: usize) -> Self {
-        debug_assert!(pos < 128);
-        1u128 << pos
-    }
-    #[inline]
-    fn clear(&mut self) {
-        *self = 0;
-    }
-    #[inline]
-    fn insert(&mut self, pos: usize) {
-        debug_assert!(pos < 128);
-        *self |= 1u128 << pos;
-    }
-    #[inline]
-    fn contains(&self, pos: usize) -> bool {
-        debug_assert!(pos < 128);
-        (*self >> pos) & 1 == 1
-    }
-    #[inline]
-    fn is_empty(&self) -> bool {
-        *self == 0
-    }
-    #[inline]
-    fn union_in_place(&mut self, other: &Self) {
-        *self |= *other;
-    }
-    #[inline]
-    fn singleton_pos(&self) -> usize {
-        debug_assert_eq!(
-            self.count_ones(),
-            1,
-            "Invoked 'singleton_pos' on a set that contains more than one element."
-        );
-        self.trailing_zeros() as usize
-    }
-    #[inline]
-    fn iter(&self) -> impl Iterator<Item = usize> + '_ {
-        let mut bits = *self;
-        std::iter::from_fn(move || {
-            if bits == 0 {
-                None
-            } else {
-                let pos = bits.trailing_zeros() as usize;
-                bits &= bits - 1;
-                Some(pos)
-            }
-        })
-    }
-}
 
 /// One position in the pattern. Each atom consumes exactly one input
 /// byte, *except* [`Atom::Any`] which can self-loop on input or be
